The newest working bmr,bmi and etc.

+++ src/application/BmiCalculation.java
@@ -1,7 +1,7 @@
 package application;
 
 public class BmiCalculation {
-	private int bmiValue;
+	private double bmiValue;
 
 	public BmiCalculation(int userWeight, int userHeight) {
 		bmiValue = (int) (userWeight / Math.pow(userHeight, 2)/100);
